Moving package from utils to modelstesting.
